Lagrer de ulike aktivitetsdagene som egne typer på kafka

+++ dto-kafka/src/main/kotlin/no/nav/aap/dto/kafka/ForrigeMottakereKafkaDto.kt
@@ -69,7 +69,7 @@ data class ForrigeMottakereKafkaDto(
     data class Paragraf_11_19_3_leddKafkaDto(
         val dato: LocalDate,
         val grunnlagsfaktor: Double,
-        val grunnbeløp: Double,
+        val grunnbeløp: Int,
         val grunnlag: Double
     )
